Provide an alternative fontsrv renderer.
This fontsrv renderer does a more thorough job at asking X11 and libxft about the metrics
of the fonts used. It seems to work well with Source Code Pro and Source Sans Pro respectively
on my 4k monitor and my 2880x1280 screen on the laptop.

It also includes a hinting change by Ethan Burns so we render fonts with no hinting and let
the system itself decide on hinting.